Move the profile creation signal so that it's an addressable function.
